cargo: remove features
The justification for them existing was that cargo adds in binary-only
deps to mixed bin/lib crates, which inflates dependency trees and
overall just sucks.

While this is true, that doesn't apply to dev-dependencies, which don't
get propagated, but which are available to example code. By removing the
myriad of features and moving those dependencies into `dev-dependencies`
we make developers' lives easier (it's easier to run examples) and don't
actually lose anything.

Moreover, this also allows to test all code with no need for
`--all-features`.